+++ packages/bonsoir_android/android/src/main/kotlin/fr/skyost/bonsoir/BonsoirService.kt
@@ -19,13 +19,22 @@ private fun NsdServiceInfo.getHostnameCompat(): String? {
     }
 }
 
+@SuppressLint("NewApi")
+private fun NsdServiceInfo.getHostAddressesCompat(): List<String> {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+        hostAddresses.map { it.hostAddress }
+    } else {
+        listOfNotNull(host?.hostAddress)
+    }
+}
+
 /**
  * Represents a Bonsoir service.
  *
  * @param name The service name.
  * @param type The service type.
  * @param port The service port.
- * @param hostAddress The service host address.
+ * @param hostAddresses The service host addresses.
  * @param hostname The service mDNS hostname.
  * @param attributes The service attributes.
  */
@@ -33,7 +42,7 @@ data class BonsoirService(
     var name: String,
     val type: String,
     var port: Int,
-    var hostAddress: String?,
+    var hostAddresses: List<String>,
     var hostname: String?,
     var attributes: MutableMap<String, String>
 ) {
@@ -46,7 +55,7 @@ data class BonsoirService(
         service.serviceName,
         if (service.serviceType.endsWith(".")) service.serviceType.substring(0, service.serviceType.length - 1) else service.serviceType,
         service.port,
-        service.host?.hostAddress,
+        service.getHostAddressesCompat(),
         service.getHostnameCompat(),
         hashMapOf(),
     ) {
@@ -67,7 +76,7 @@ data class BonsoirService(
             "${prefix}name" to name,
             "${prefix}type" to type,
             "${prefix}port" to port,
-            "${prefix}hostAddress" to hostAddress,
+            "${prefix}hostAddresses" to hostAddresses,
             "${prefix}hostname" to hostname,
             "${prefix}attributes" to attributes,
         )
@@ -84,8 +93,12 @@ data class BonsoirService(
             serviceType = type
             port = this@BonsoirService.port
         }
-        if (hostAddress != null) {
-            service.host = InetAddress.getByName(hostAddress)
+        if (hostAddresses.isNotEmpty()) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+                service.hostAddresses = hostAddresses.map { InetAddress.getByName(it) }
+            } else {
+                service.host = InetAddress.getByName(hostAddresses.first())
+            }
         }
         for (entry in attributes.entries) {
             service.setAttribute(entry.key, entry.value)