+++ packages/bonsoir_darwin/darwin/bonsoir_darwin/Sources/bonsoir_darwin/SwiftBonsoirPlugin.swift
@@ -41,9 +41,9 @@ public class SwiftBonsoirPlugin: NSObject, FlutterPlugin {
         let id = arguments["id"] as! Int
         switch call.method {
         case "broadcast.initialize":
-            let service = BonsoirService(name: arguments["service.name"] as! String, type: arguments["service.type"] as! String, port: arguments["service.port"] as! Int, hostAddress: nil, attributes: arguments["service.attributes"] as! [String: String])
-            if let hostAddress = arguments["service.hostAddress"] as? String? {
-                service.hostAddress = hostAddress
+            let service = BonsoirService(name: arguments["service.name"] as! String, type: arguments["service.type"] as! String, port: arguments["service.port"] as! Int, hostAddresses: [], attributes: arguments["service.attributes"] as! [String: String])
+            if let hostAddresses = arguments["service.hostAddresses"] as? [String] {
+                service.hostAddresses = hostAddresses
             }
             if let hostname = arguments["service.hostname"] as? String? {
                 service.hostname = hostname