+++ packages/bonsoir_windows/windows/bonsoir_service.h
@@ -3,6 +3,7 @@
 #include <flutter/encodable_value.h>
 
 #include <optional>
+#include <vector>
 
 using namespace flutter;
 
@@ -12,7 +13,7 @@ namespace bonsoir_windows {
     std::string name;
     std::string type;
     int port;
-    std::optional<std::string> hostAddress;
+    std::vector<std::string> hostAddresses;
     std::optional<std::string> hostname;
     std::map<std::string, std::string> attributes;
 
@@ -20,7 +21,7 @@ namespace bonsoir_windows {
       std::string _name,
       std::string _type,
       int _port,
-      std::optional<std::string> hostAddress,
+      std::vector<std::string> hostAddresses,
       std::optional<std::string> hostname,
       std::map<std::string, std::string> _attributes
     );