+++ packages/bonsoir_android/CHANGELOG.md
@@ -75,7 +75,7 @@
 
 ## 4.0.0
 
-* `service.ip` is now `service.hostAddress`.
+* `service.ip` is now `service.hostAddresses`.
 
 ## 3.0.0+1